Get diversity table

Compute a table with various diversity indices from a species occurrence dataframe.

Usage

get_diversity_table(
  df,
  cam_col,
  spp_col,
  count_col = NULL,
  keep_all_levels = TRUE
)

Arguments

df

The dataframe

cam_col

Name of the column containing cameras names

spp_col

Name of the column containing species names

count_col

Name of the column containing species counts (optional)

keep_all_levels

Should we keep all camera levels even those that have no rows in df? (Checked only if cam_col is a factor)

Value

A dataframe with one row per camera-species combination. For each camera-species, we have the count (column count) and proportion (column prop) of individuals of each species seen at this camera.

Examples

df <- data.frame(species = c("zebra", "cat", "cat", "cow", "cow", NA, NA),
                 type = c("animal", "animal", "animal", "animal", "animal", "human", "blank"),
                 camera = c("C1", "C1", "C1", "C1", "C2", "C2", "C3"),
                 count = c(1, 1, 1, 50, 3, 1, NA))
get_diversity_table(df,
                    cam_col = "camera",
                    spp_col = "species",
                    count_col = "count")
#>   camera species count       prop
#> 1     C1     cat     2 0.03773585
#> 2     C1     cow    50 0.94339623
#> 3     C1   zebra     1 0.01886792
#> 4     C2     cow     3 0.75000000
#> 5     C2    <NA>     1 0.25000000
#> 6     C3    <NA>    NA         NA