Summarize species

Summarize species information from a data table

Usage

summarize_species(
  df,
  species_col,
  cam_col = NULL,
  obs_col = NULL,
  count_col = NULL,
  ncam = NULL,
  NA_count_placeholder = NA
)

Arguments

df

the observation dataframe to summarize

species_col

Name of the species column

cam_col

Name of the column containing camera codes (optional)

obs_col

Name of the observation type column (optional)

count_col

Name of the column containing species count (optional)

ncam

Number of cameras to take into account when computing the proportion of cameras the species was ween on. If NULL, defaults to the number of cameras present in the df.

NA_count_placeholder

Value with which to replace NAs present in the column containing counts. Defaults to NA (i.e. values are not replaced)

Value

A table summarizing species information with the following columns:

Species (named like species_col): species identity (same as the species_col input column)

Observation type (present only if obs_col is not NULL and named like obs_col): observation type (same as the obs_col input column)

n_sightings: number of rows where the species was photographed.

n_individuals: count of individuals observed on all pictures (using the input count_col column). If count_col is NULL, it contains the same values as n_sightings. If there are NAs in the input count_col, they will propagate in n_individuals (unless a value is specified in NA_count_placeholder).

n_cameras (present only if cam_col is not NULL) : the number of cameras the species was seen on.

prop_cam (present only if cam_col is not NULL): the proportion of cameras the species was seen on. If ncam is provided, then it uses ncam as the total number of cameras.

Examples

df <- data.frame(species = c("zebra", "cat", "cat", "cow", NA, NA),
                 type = c("animal", "animal", "animal", "animal", "human", "blank"),
                 camera = c("C1", "C1", "C2", "C3", "C3", "C4"),
                 count = c(1, 1, 3, 50, 1, NA))
res <- summarize_species(df, 
                         species_col = "species", cam_col = "camera",
                         obs_col = "type",
                         count_col = "count",
                         NA_count_placeholder = 1)