Get non-shared cameras between two dataframes

Using two dataframes in input, determine which cameras are in one of the tables but not in the other one.

Usage

get_cameras_not_in(dfrecords, dfcameras, cam_col_dfrec, cam_col_dfcam)

Arguments

dfrecords

records dataframe

dfcameras

cameras dataframe

cam_col_dfrec

name of the cameras column in the records dataframe

cam_col_dfcam

name of the cameras column in the cameras dataframe

Value

A named list with two components

$not_in_records: cameras from dfcameras that are not in dfrecords. If all cameras from dfcameras are in dfrecords, this is a character vector of length zero.

$not_in_cameras: cameras from dfrecords that are not in dfcameras If all cameras from dfrecords are in dfcameras, this is a character vector of length zero.

Examples

dfrecords <- data.frame(camID = letters[2:7])
dfcam <- data.frame(cameras = letters[1:5])
get_cameras_not_in(dfrecords = dfrecords, 
                   dfcameras = dfcam,
                   cam_col_dfrec = "camID",
                   cam_col_dfcam = "cameras")
#> $not_in_records
#> [1] "a"
#> 
#> $not_in_cameras
#> [1] "f" "g"
#>