Get diversity indices

From a summary table of camera/species, return richness, Shannon and Simpson diversity indices.

Usage

get_diversity_indices(
  count_df,
  spp_col,
  cam_col,
  count_col = "count",
  prop_col = "prop"
)

Arguments

count_df

Dataframe summarizing counts per species. Can be computed using the records dataframe with the get_diversity_table function.

spp_col

Name of the column containing species names

cam_col

Name of the column containing cameras names

count_col

Name of the column containing species counts

prop_col

Name of the column containing species proportions

Value

A dataframe with one row per camera summarizing diversity indices.

Details

Richness is computed as the number of the different species seen at the camera.

Shannon index is computed as: $$-\sum p_i ln(p_i)$$ Where p_i represents the proportion of the abundance of species i at a camera over the the total number of individuals of all species seen at this camera. It ranges between 0 and +infinity, zero indicating the lowest diversity.

Simpson index is computed as: $$(\sum n_i (n_i - 1))/(N (N-1))$$ Where n_i represents the abundance of species i at a camera over the total number of individuals of all species seen at this camera. It ranges between 0 and 1, one indicating the lowest diversity.

Examples

countdf <- data.frame(camera = c("C1", "C1", "C1", "C2",
                                 "C3", "C3", "C3"),
                      species = c("cat", "cow", "rabbit",
                                  "cat", "cat", "cow", "rabbit"),
                      count = c(30, 30, 30, 30, 88, 1, 1),
                      prop = c(1/3, 1/3, 1/3, 1, 88/90, 1/90, 1/90))
get_diversity_indices(countdf, 
                      spp_col = "species", cam_col = "camera")
#>   camera richness   shannon   simpson
#> 1     C1        3 1.0986123 0.3258427
#> 2     C2        1 0.0000000 1.0000000
#> 3     C3        3 0.1219692 0.9558052